refactor|basic-test-snapshot
