+++ src/component/basic/index.js
@@ -1,0 +1,12 @@
+import React from 'react';
+
+export const Basic = () => {
+  return (
+    <div >
+      <h1> Basic Test</h1>
+      <p> This is a basic Test Component</p>
+    </div>
+  );
+}
+
+export default Basic;
